+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased changes
 
--
+### 🐞 Fixes
+
+- Fix usage of more than possessive quantifier in the same regex.
 
 ## Released changes